% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurver_aggr.R
\name{epicurver_aggr}
\alias{epicurver_aggr}
\title{Epicurver for aggregated data}
\usage{
epicurver_aggr(data, xvar, yvar, group = NULL, col_scale = "qualitative",
  border = NA, xtitle = "Time", legend_title = group,
  x_axis_ticks = c(NULL, "days", "2 days", "weeks", "2 weeks", "months",
  "years"), x_axis_limits = NULL)
}
\arguments{
\item{data}{Your data frame.}

\item{xvar}{Time variable, given without quotes.}

\item{yvar}{A variable with the number of cases for each time unit, given without quotes.}

\item{group}{If applicable, a grouping variable, given without quotes. Defaults to one group. 
In case of groupings, works up to 7 groups with ECDC qualitative colours.}

\item{col_scale}{Colour scale for the bars, defaults to qualitative colour scale; select one of the following: "qualitative",
"green", "blue", "red".}

\item{border}{Bar border colour, defaults to no border (NA).}

\item{xtitle}{x-axis title.}

\item{x_axis_ticks}{Tick marks for x-axis. Use ONLY if x-axis uses dates or months; select one of the following: "days", "2 days", 
"weeks", "2 weeks", "months", "years". Defaults to "days" with days and to numbers with numeric weeks or months. Selecting "weeks" prints only 
mondays for x-axis ticks. In case of "months" labelled automatically with abbreviated 
month names.}

\item{x_axis_limits}{x-axis limits, for dates use character dates as c("2017-01-01", "2017-12-31"), otherwise a numeric vector lenght of two.}

\item{Legend_title}{Legend title.}
}
\description{
This function draws an epicurve for aggregated data. Draws a bar for 
each time unit (usually week, month, year, but also dates are possible). 
Works with data that has been pre-aggregated per time 
unit, and with a continuous x-axis (e.g. months or weeks as numbers). 
Allows using facet_grid() for quick and dirty month/week epicurve spanning over 
several years.
}
\examples{
# Create dummy data
mydat <- data.frame(ID = c(seq(1,24,1)),
Case = c(sample(10:50, 24, replace = FALSE)),
Month = rep(c(seq(1,12,1)),2),
Gender = c(rep("F",12), rep("M", 12)))

# Plot one epicurve without any groupings
epicurver_aggr(data = mydat[1:12,], xvar = Month, yvar = Case, 
xtitle = "Month of onset")

# Plot by two level grouping (e.g. Gender), with tick marks as months
epicurver_aggr(data = mydat, xvar = Month, yvar = Case, 
xtitle = "Month of onset", group = Gender, x_axis_ticks = "months")
}
\author{
Tommi Karki
}
\keyword{epicurve}
\keyword{epicurver,}
