% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.R
\name{ggbar}
\alias{ggbar}
\title{Simple bar chart with ECDC colours}
\usage{
ggbar(data, xvar, group = NULL, col_scale = c("qualitative", "green",
  "blue", "red"), xtitle = xvar, ytitle = "count", legend_title = group)
}
\arguments{
\item{data}{Your data frame}

\item{group}{vector in your dataframe, given without quotes}

\item{col_scale}{colour scale, defaults to "qualitative"}

\item{xtitle}{x axis title, defaults to x}

\item{ytitle}{y axis title, defaults to "count"}

\item{legend_title}{legend title, if applicable defaults to group}

\item{x}{vector in your dataframe, given without quotes}
}
\description{
Somewhat pipe (dplyr) friendly bar chart. Allows a stacked bar chart with groupings.
}
\examples{
Create dummy data
mydat <- data.frame(ID = c(seq(1,10,1)),
Gender = c(rep(c("F", "M"),5)),
AgeGroup = c(rep(c("0-18", "18-65", "65+"),3), "65+"),
Case = c(1,1,1,0,0,1,1,1,1,1),
DateOfOnset = as.Date(c("2017-06-11", "2017-06-11", 
                       "2017-06-11", NA, NA, "2017-06-10", 
                       "2017-06-14", "2017-06-14",
                       "2017-06-19", "2017-06-19")),
Month = c(sample(c(3:6),5, replace = TRUE), sample(c(1:12),5)),
Week = c(sample(c(10:12),5, replace = TRUE), sample(c(1:53),5)))

bar(mydat, Gender)

#Stacked bar by AgeGroup
bar(mydat, Gender, group = AgeGroup)

}
\seealso{
ggplot2
}
\author{
Tommi Karki
}
\keyword{bar}
\keyword{chart}
